requirements.txt
cython
numpy
tqdm
opencv-contrib-python
pillow
absl-py
tensorflow
matplotlib
nlib3
pandas
# Data preprocessing relations
scipy


# Model diagram output
# pydot
# pydotplus
# graphviz


# ResNet-RS trained model
# git+https://github.com/sebastian-sz/resnet-rs-keras@main


# pypi
wheel
twine
